Add files via upload
There are two files in this folder. One is coded to be loading page and the other for the dice game. The loading page has a button which connects the loading page to the dice game. So click on DICEGAME.html and click the button to load up the game.  The coding for dice game is in the file customization.html.